implement retry logic for deb_chroot
A built package might be a dependency of another package. This might
make it more difficult to properly build all the packages.
For that, a simple (but a little heavy) way to handle that is this loop:

while all the packages are built or no new packages could be built:
1) build all the packages we can, making errors not fatal
2) create a small local repository, used by the chroot for the next
iteration